Use 'SPARK_DRIVER_LOG_URL_' as env prefix for getting driver log urls by default
